# Simplify x-element.