--- a/demo/demo-element.js
+++ b/demo/demo-element.js
@@ -1,57 +1,40 @@
-import XElementBasic from '../x-element-basic.js';
+import XElement from '../x-element.js';
 
-class DemoElement extends XElementBasic {
-  static template() {
-    return ({ reflected }) => `
-      <style>
-        :host {
-          display: block;
-          width: 200px;
-        }
-
-        :host([reflected]) {
-          background-color: yellow;
-        }
-
-        :host([boolean-value]) {
-          font-weight: bold;
-        }
-      </style>
-      <div id="demo">${reflected}</div>
-    `;
+class DemoElement extends XElement {
+  static get properties() {
+    return {
+      reflected: {
+        type: String,
+        reflect: true,
+      },
+      booleanValue: {
+        type: Boolean,
+        reflect: true,
+        initial: true,
+      },
+    };
   }
 
-  static get observedAttributes() {
-    return ['reflected', 'booleanValue'];
-  }
-
-  set reflected(value) {
-    if (value) {
-      this.setAttribute('reflected', value);
-    } else {
-      this.removeAttribute('reflected');
-    }
-  }
-
-  get reflected() {
-    return this.getAttribute('reflected');
-  }
-
-  set booleanValue(value) {
-    if (value) {
-      this.setAttribute('boolean-value', '');
-    } else {
-      this.removeAttribute('boolean-value');
-    }
-  }
-
-  get booleanValue() {
-    return this.getAttribute('boolean-value');
-  }
-
-  connectedCallback() {
-    super.connectedCallback();
-    this.booleanValue = true;
+  static template(html) {
+    return ({ reflected }) => {
+      return html`
+        <style>
+          :host {
+            display: block;
+            width: 200px;
+          }
+          
+          :host([reflected]) {
+            background-color: yellow;
+          }
+          
+          :host([boolean-value]) {
+            font-weight: bold;
+          }
+        </style>
+        <div id="demo">${reflected}</div>
+      `;
+    };
   }
 }