--- a/x-element-basic.js
+++ b/x-element-basic.js
@@ -1,6 +1,0 @@
-/**
- * Implements a naive template rendering system and a few helpers.
- */
-import ElementMixin from './mixins/element-mixin.js';
-
-export default ElementMixin(HTMLElement);