… CHANGELOG.md
… README.md
… SPEC.md
… demo/index.html
… etc/graph.js
… mixins/element-mixin.js
… mixins/listeners-mixin.js
… mixins/lit-html-mixin.js
… mixins/properties-mixin.js
… mixins/property-effects-mixin.js
… package.json
… test/fixture-element-attr-binding.js
… test/fixture-element-attr-reflection.js
… test/fixture-element-basic.js
… test/fixture-element-computed-properties.js
… test/fixture-element-listeners.js
… test/fixture-element-observed-properties.js
… test/fixture-element-read-only-properties.js
… test/fixture-element-scratch.js
… test/fixture-element-upgrade.js
… test/index.html
… test/test-analysis-errors.html
… test/test-analysis-errors.js
… test/test-attr-binding.html
… test/test-attr-binding.js
… test/test-attr-reflection.html
… test/test-attr-reflection.js
… test/test-attribute-changed-errors.html
… test/test-basic-properties.html
… test/test-basic-properties.js