Concept branch — the goal here is to reimagine `x-element` based on our learnings from using it in hundreds of components in a complex SPA over the last year.

Significant changes:

* Static analysis is done up-front, once.
* Anything that can be validated statically throws errors _except for interface shadowing_ which changes as browsers update — the best we can do is `console.warn` in this case.
* _Almost all_ of the XElement interface is static (though we provide trivial convenience wrappers). Note that `render` is the only non-trivial instance method.
* Static interface is ready for private (`#`) static fields when they're ready.
* `observer` >> `observe` is now a function (proper context is maintained if it's a constructor method).
* `computed` >> `input` + `compute`. `compute` is now a function (proper context is maintained if it's a constructor method). `input` is an array of property names which will be resolved to arguments at runtime and passed to `compute`.
* `value` is now `initial` and `default`. `initial` is used if initial value is nullish. `default` is used if property value is ever nullish.
* Type coercion is no longer done on set.
* Errors are thrown when you set a value with an incorrect type in the runtime.
* Concept of `readOnly` is first-class.
* Concept of `internal` is first-class.
* Internal setting/getting and read-only setting can be done via the `.internal` proxy.
* Computed properties are now lazily evaluated.
* Observation, reflection, and computation are no longer synchronous. Importantly, this allows us to set properties one-by-one and then get a single `compute` at the end.
* Concept of a "render root" is now supported via `createRenderRoot`.
* Initial template function is now cached on each host (vs defining a new function on each render).
* We always allow computation of computed properties before any arguments are defined. Previously, we had thought we might want to defer initial compute until at least one argument was defined.
* Swapped the order of values in `observe` callbacks to be `value, oldValue` — it was previously `oldValue, value`.
* Order of operations are now compute, reflect, render, then observe.
* Listener logic has been improved and interface is now a function, not a string (proper context is maintained if it's a constructor method).
* Coercion of `null` and `undefined` to `''` is removed, the underlying templating engine takes care of this for us.
* The notion of a render proxy is gone. Previously we had only used it to convert `null` and `undefined` to the empty string. However, that is handled by the templating engine now.
* Deserialization of anything other than `[Boolean, String, Number]` now throws a halting error (e.g., setting an attribute for a property that is of `type: Object`). Previously, we basically did nothing, especially for custom classes.