@theengineear 2020-05-26
@klebba — mind taking a look when you have the chance? Also happy to walk you through the changes. I think you'll like the general direction here.