@theengineear 2020-05-27
@klebba — I left a couple things in limbo here. Please hold off on looking at the code until I get the chance to fix.