@klebba 2020-05-29
I tried swapping this in for a feature I'm working and might have caught an issue; my observer functions are not being. Here's my property block:
```
  static get properties() {
    return {
      targetRect: {
        type: DOMRect,
      },
      open: {
        type: Boolean,
        reflect: true,
        observer: this.observeOpen,
      },
      _parentElement: {
        type: HTMLElement,
      },
      _cssProperties: {
        type: Object,
        dependencies: ['_parentElement', 'targetRect', 'open'],
        resolver: this.computeCSSProperties,
        observer: this.observeCSSProperties,
      },
    };
  }
```

My demo page has markup like:

```
<x-inspector debug open .targetRect="${rect1}">[...]</x-inspector>
```

So some of the values need to be read immediately following element upgrade. Also I just grabbed whatever I saw in the PR for `x-element.js` a few minutes ago so maybe it's not ready for this yet, but I thought I'd give it a try because I think the change in order for render/reflect etc might be helpful for this feature