@klebba 2020-06-03
```
static async observePage(host, current, last) {
```

This throw an error:
```
Uncaught Error: Unexpected value for "XPages.properties.page.observer" (expected Function).
```

If I remove the `async` it appears to work