@klebba 2020-06-03
`Uncaught Error: Unexpected value for "XPages.properties.currentElement" (expected HTMLElement).`

Would be nice to say `got ${value}` here as well