@klebba 2020-06-03
```
Error: Property "XPages.properties.currentElement" has resolver (cannot be set).
```

Might read better as:

```
Error: Property "XPages.properties.currentElement" is read only. (defined by resolver ${resolverName}).
```

Or similar