@klebba 2020-06-03
```
observer: this.observeModel,
```

Does not throw even though this function does not exist