@klebba 2020-06-03
Here's an issue that seems like a bug:

```
    model: {
        type: Object,
        observer: this.observeModel,
      },
      currentElement: {
        observer: this.observerCurrentElement,
      },
      binding: {
        type: Boolean,
        internal: true,
        dependencies: ['currentElement', 'model'],
        resolver: this.computeBinding,
      },
```

The resolver function for `binding` is never called even though I'm able to confirm that both observers (`this.observerModel` and `this.observerCurrentElement`) are being called with the values I would expect. Note: I'm trying to port `x-pages`