@theengineear 2020-06-03
>If I remove the async it appears to work

Thanks, need to check for either `AsyncFunction` or `Function`.