@theengineear 2020-06-03
>observer: this.observeModel,
>Does not throw even though this function does not exist

Yah, I've been loose about `undefined`. I'll change it to a `Reflect.has` check instead of `!== undefined`.