@theengineear 2020-06-03
>The resolver function for binding is never called even though I'm able to confirm that both observers (this.observerModel and this.observerCurrentElement) are being called with the values I would expect. Note: I'm trying to port x-pages

Are you sure it's `this.computeBinding,` and not `this.resolveBinding`? I'm wondering if the loose `undefined` check was just problematic.