@theengineear 2020-07-06
**CAVEAT** these tests are super nascent and we should not draw any final conclusions just yet!

A _really_ quick, cursory study of performance yields the following:

```
[THIS BRANCH]
average definition time (count = 100000): 0.028538199957692995 ms
average instantiation time (count = 100000): 0.006306099989451468 ms
average assignment time (count = 100000): 0.000526649983949028 ms
average update time (count = 100000): 0.0028123500145738945 ms

[MASTER]
average definition time (count = 100000): 0.013228350005229005 ms
average instantiation time (count = 100000): 0.01043335000926163 ms
average assignment time (count = 100000): 0.0005833499878644943 ms
average update time (count = 100000): 0.0028958999877795578 ms
```

I decided not to share the code since this is just an initial glance — we'd need more nuanced tests to really dig in. However, here's a quick analysis of the findings:

* As expected, definition time analysis takes longer now, ~2x slower to define an element. As a gut check, if we were to import and define _one thousand_ elements, it would still be just a fraction of our overall application boot time.
* Pertaining to the original thread — assignment and update remain nearly unchanged. I.e., there seems to be no bottleneck around type checking.
* As hoped for, there seems to be a discernible improvement in instantiation time — ~1.5 faster to instantiate an element.